Island: Rename cc/resources/auth/registration.py -> cc/resources/auth/register.py
