+++ monkey/monkey_island/cc/resources/auth/register.py
